Add OverlayController
